chore: fix local dev
