+++ packages/compass-collections-ddl/electron/renderer/mocks/collection-store.js
